+++ CHANGELOG.md
@@ -10,6 +10,7 @@ To see unreleased changes, please see the [CHANGELOG on the master branch](https
 
 ### Added
 
+* `ErrorInfo` got optional `root_module` field.
 * `py.typed` file for PEP-561 compatibility
 * docs: "Supported standards" page